Add optional `duration` to `send_task_to_worker`
In some cases we may have already computed the `duration`, so provide an
optional parameter to supply this if it is available. Otherwise go ahead
and compute it here.

Since we already compute the `duration` in
`transition_waiting_processing`, go ahead and pass that along to
`send_task_to_worker` to avoid needing to compute it twice.