Optimize `send_task_to_worker` (#4331)
* Move validation block with dependency handling

As both of these branches check that `deps` is non-trivial, we can skip
the check in the second branch if we include it within the branch where
the check already happened. Should simplify the work needed for this
check.

* Group `WorkerState` annotation with others

* Annotate `msg` and `deps` as `dict` and `set`

* Use `_key` attribute from `TaskState`

* Take `TaskState` object in `send_task_to_worker`

All that is needed in `send_task_to_worker` can already be gotten from
the `TaskState` object, which we retrieve in
`transition_waiting_processing` anyways. So just pass the `TaskState`
object directly instead of the `key`, which is also in the `TaskState`
object.

* Add optional `duration` to `send_task_to_worker`

In some cases we may have already computed the `duration`, so provide an
optional parameter to supply this if it is available. Otherwise go ahead
and compute it here.

Since we already compute the `duration` in
`transition_waiting_processing`, go ahead and pass that along to
`send_task_to_worker` to avoid needing to compute it twice.