Optionally use offload executor in worker (#4307)
Normally we have one separate thread pool executor for deserializing data
and another thread pool executor in a worker for execution.
Sometimes this presents a problem because some data types want to be
used in the thread in which they were created.  One example of this is
TensorFlow graphs, but there are others.

One way to resolve this is to reuse the same executor in both
situations, and ensure that it has only one thread.  This means that
execution and deserialization will block each other (not great) but that
user data will always be operated on in one thread only.

Co-authored-by: James Bourbeau <jrbourbeau@gmail.com>