Just use `msg.get("key")`
Instead of checking for `"key"` in `msg`, just use `.get(...)` to try
and retrieve the `key`. If it fails, we get `None`, which we can check
and ignore quickly. If not, we get the value and can proceed to use it.
Should avoid needing to check for the presence of `"key"` in `msg`
twice.