Use `dict.get(...)` instead of catching error
This ends up being a bit faster than catching the error (especially in
Cython/C where it knows this error can just be cleared instead of
handled). Plus the comparison to `None` is fast in both (in Cython/C
this is a simple pointer value comparison similar to comparing to
`NULL`).