Use `not not` instead of `bool`
This winds up being a faster way to test truthiness (even in pure
Python). Plus when Cython looks at this code, it can skip constructing a
Python `bool` object and simply operate on the results in C directly.

```python
In [1]: s0 = set()

In [2]: s1 = {1, 2, 3}

In [3]: %timeit bool(s0)
66.1 ns ± 0.123 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [4]: %timeit bool(s1)
70.5 ns ± 0.203 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [5]: %timeit (not not s0)
22.6 ns ± 0.25 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [6]: %timeit (not not s1)
28.3 ns ± 1.35 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```