Annotation of some comm related methods in the Scheduler (#4341)
* Annotate `stream_comms` in `worker_send`

Provides type annotation for `stream_comms` in the form of a locally
stored variable in `worker_send`. This should allow the Python C API to
used for lookups when the scheduler is Cythonized.

* Just use `msg.get("key")`

Instead of checking for `"key"` in `msg`, just use `.get(...)` to try
and retrieve the `key`. If it fails, we get `None`, which we can check
and ignore quickly. If not, we get the value and can proceed to use it.
Should avoid needing to check for the presence of `"key"` in `msg`
twice.

* Store `self.tasks` to locally annotated variable

Should allow more efficient access of `tasks` in Cython as it is known
to be a `dict`.

* Annotate `client_keys` as a `list` in `report`

* Set locally annotated var with `self.client_comms`

Should allow more efficient access of `client_comms` in Cython as it is
known to be a `dict`.

* Annotate `k` as `str`

* Use `dict.get(...)` instead of catching error

This ends up being a bit faster than catching the error (especially in
Cython/C where it knows this error can just be cleared instead of
handled). Plus the comparison to `None` is fast in both (in Cython/C
this is a simple pointer value comparison similar to comparing to
`NULL`).

* Annotate `report_on_key`'s function arguments

* Use `!=` to emulate `xor` in Python

* Store `self.tasks` to locally annotated variable

Should allow more efficient access of `tasks` in Cython as it is known
to be a `dict`.

* Use `dict.get(...)` instead of catching error

This ends up being a bit faster than catching the error (especially in
Cython/C where it knows this error can just be cleared instead of
handled). Plus the comparison to `None` is fast in both (in Cython/C
this is a simple pointer value comparison similar to comparing to
`NULL`).