Don't break actors
This was an initial attempted fix for a test failure in
`test_worker.py::test_clean_nbytes` that turns out to be unnecessary.
On top of that, it breaks actors, so it shouldn't be in here.

For reference, the value of actor tasks goes in a separate `self.actors`
dict, not `self.data`