Some common optimizations across transitions (#4348)
* Overwrite `s` in non-`None` `valid_workers` case

* Annotate some variables in `_propagate_forgotten`

* Type `recommendations` as `dict`

* Assign `task` to locally typed variable

* Use `not not` for faster, C amenable truthiness

As shown before `not not` is faster than `bool`. Plus Cython is able to
use the result from `not not` directly in C without having to convert
back to a Python object as is the case with `bool`. Thus resulting in
faster comparisons in the C generated by Cython.

* Annotate local variables in `_add_to_memory`

* Coerce `ts._dependents` to a `list`

This is what `sorted` will do anyways. Plus it simplifies the logic on
Cython later when constructing the `for`-loop over `deps`.

* Use faster `state` attribute access on `TaskState`