Use `TaskState` w/`report_on_key` in Scheduler
As `report_on_key` needs the `TaskState` regardless and will just fetch
it if given the `key`, go ahead and just pass the `TaskState` object
instead of the `key` when available. After all `TaskState` is a typed
object that has fast attribute access. So we can fetch anything else we
need (including the key) anyways.