Merge `assert` handling into argument handling
As we already need to go through and handle `ts is None` and the second
branch should already be `key is None` (and is easy enough to check), go
ahead and move the `assert` into the `else` case at the end. All of
these comparisons are really fast pointer comparisons that we were
already doing before more-or-less. So finishing it out with an `assert
False`, doesn't really cost us anything and is more optimized for both
pure Python and Cython.