Optimizations in notable functions used by transitions (#4351)
* Type `w` as `str` to match `ws._address`

* Use `TaskState` w/`report_on_key` in Scheduler

As `report_on_key` needs the `TaskState` regardless and will just fetch
it if given the `key`, go ahead and just pass the `TaskState` object
instead of the `key` when available. After all `TaskState` is a typed
object that has fast attribute access. So we can fetch anything else we
need (including the key) anyways.

* Inline `self._nbytes` usage in `get_nbytes`

* Make `sum` of `list` comprehension a `for`-loop

Avoids some extra boilerplate that Cython would generate to build the
`list` and then loop through it with `sum`. Instead this builds a single
`for`-loop over `ts._dependencies` and handles the addition in C. Should
improve the performance here a bit.

* Add `stack_time` first

This is slightly easier to read. Also Cython will translate this into
doing some work first before checking for zero in `self.bandwidth`.

* Annotate and compute times in C

Make sure that variables like `self.bandwidth` are cast to C variables.
Perform computations with these in straight C.

* Merge `assert` handling into argument handling

As we already need to go through and handle `ts is None` and the second
branch should already be `key is None` (and is easy enough to check), go
ahead and move the `assert` into the `else` case at the end. All of
these comparisons are really fast pointer comparisons that we were
already doing before more-or-less. So finishing it out with an `assert
False`, doesn't really cost us anything and is more optimized for both
pure Python and Cython.

* Break out 2nd handling of `ts is None`

As the `key` may be missing from `TaskState`, in which case `ts` will
still be `None` after trying to retrieve it from `self.tasks`, handle
that `if` as well, but handle it like the other `report` conditions and
not separately. Simplifies the code a bit and improves readability.