Use `SortedDict` for `self.idle`
Instead of using `SortedSet` for `self.idle`, use `SortedDict`. This
avoids needing `operator` to retrieve `WorkerState.address`, which we
can retrieve on our own (and do so faster!). Plus `SortedDict` inherits
from `dict`, which allows us to use some Python C API functions on it.
Something that `SortedSet` does not.