Cast `SortedDict` to `dict` when viewing
As `SortedDict` inherits from `dict`, we can ask Cython to cast
`SortedDict` to `dict`. This is helpful as Cython will make use of
efficient Python C APIs for `dict`s if it knows the value is typed that
way. So should improve performance of things like item retrieval, checks
for inclusion of keys, iteration, etc..

That said, it's worth noting that operations that manipulate the
contents of the `SortedDict` should not use the Python C API as
`SortedDict` overrides all of these methods to provide the `sorted`
behavior we expect. So when adding or removing values, we use the actual
`SortedDict` object not the cast representation.