Cast `SortedDict`s to `dict`s in a few key places & other minor changes (#4355)
* Use `WorkerState` `address` attribute

Allows for faster access via C as opposed to via Python.

* Annotate `n_workers` and `n_tasks`

Moves the comparisons and math related to these from Python into C when
Cython is used. Otherwise has no real effect.

* Use `SortedDict` for `self.idle`

Instead of using `SortedSet` for `self.idle`, use `SortedDict`. This
avoids needing `operator` to retrieve `WorkerState.address`, which we
can retrieve on our own (and do so faster!). Plus `SortedDict` inherits
from `dict`, which allows us to use some Python C API functions on it.
Something that `SortedSet` does not.

* Cast `SortedDict` to `dict` when viewing

As `SortedDict` inherits from `dict`, we can ask Cython to cast
`SortedDict` to `dict`. This is helpful as Cython will make use of
efficient Python C APIs for `dict`s if it knows the value is typed that
way. So should improve performance of things like item retrieval, checks
for inclusion of keys, iteration, etc..

That said, it's worth noting that operations that manipulate the
contents of the `SortedDict` should not use the Python C API as
`SortedDict` overrides all of these methods to provide the `sorted`
behavior we expect. So when adding or removing values, we use the actual
`SortedDict` object not the cast representation.