Add `final` wrapper for classes w/o inheritance
By default Cython assumes that extension type objects may be subclassed.
So it creates a virtual table for all methods of the base class in the
object `struct`. This way these function pointers can be changed to
point to some other method for subclasses. However if we know a class
will not be subclassed, we can indicate that by marking it as `final`.
This allows Cython to do away with the virtual table and instead just
have a collection of functions to call on the objects. These functions
can in turn be inlined or otherwise optimized. Should be useful for the
`*State` and `Task*` objects, which are not subclassed and can benefit
from the faster function calls (like `get_nbytes`).