Optimizations of `*State` and `Task*` objects and stealing (#4358)
* Cast `workers` in `stimulus_task_finished`

* Assign `self.tasks` to `dict` typed variable

* Multiply `p` & `nc` before dividing

Avoids an extra check for zero division.

* Multiply `nc` & `avg`

This avoids a division by zero check that seems unlikely in this case
(we should have some threads).

* Delay computing `pending` until needed

* Combine comparisons of `pending`

Cython will pipeline this sort of check more effectively in C. So
leverage that behavior here.

* Simplify branch structure

* Type `tasks` as `set`

* Type `bandwidth` in `get_comm_cost`

* Type `duration` in `get_task_duration`

* Assign `set` to local variable for Cython opts

* Type `default` in `get_task_duration`

* Extract & assign `UNKNOWN_TASK_DURATION` globally

* Overwrite `duration` in `get_task_duration`

* Annotate `identity` return type as `dict`

* Type `get_nbytes` return type as `Py_ssize_t`

* Add `ccall` wrapper for `cpdef` style functions

Should allow us to make faster calls to Python functions decorated with
`ccall`.

* Add `exceptval` to tweak function exception checks

Normally the default exception check behavior from Cythonized functions
is fine. However in a few cases we need to tweak this slightly. so
import `exceptval` so that we can make those tweaks.

* Decorate all `*State` & `Task*` methods w/ `ccall`

In addition to the Python interface these methods have, this adds C
interface, which Cython will opt to use in Cython code with access to
that interface (currently just `scheduler.py`). Should speed up function
calls in these cases.

* Add `final` wrapper for classes w/o inheritance

By default Cython assumes that extension type objects may be subclassed.
So it creates a virtual table for all methods of the base class in the
object `struct`. This way these function pointers can be changed to
point to some other method for subclasses. However if we know a class
will not be subclassed, we can indicate that by marking it as `final`.
This allows Cython to do away with the virtual table and instead just
have a collection of functions to call on the objects. These functions
can in turn be inlined or otherwise optimized. Should be useful for the
`*State` and `Task*` objects, which are not subclassed and can benefit
from the faster function calls (like `get_nbytes`).

* Mark all `*State` & `Task*` objects `final`

Should eliminate virtual tables for these objects and allow inlining of
their methods.

* Add `declare` for typing variables

Ideally we would just use type annotations at the module level,
unfortunately that doesn't seem to be something Cython translates into C
typed variables. Fortunately `declare` seems to work just fine for this
purpose. So go ahead and include it so we can type module variables.
Should make it easier to use these in other functions without Cython
needing to convert them every time.

* Type module global variables with `declare`

This tells Cython that these variables have the type provided and should
be assigned the value given. By doing this, we should now have `static`
variables in C with these values, which we can use locally and Cython
will allow C operations to occur on them.

* Add `inline` to suggest inlining of functions

This can be handy when trying to suggest to the C compiler to just go
ahead an inline a function (like `get_nbytes`) to save on the cost of a
function call.

* Add `inline` to `get_nbytes`

Should encourage the compiler to just go ahead and include the code of
this method directly where it is used.

* Add `nogil` wrapper for functions

Should help us mark functions that don't require the GIL (like
`get_nbytes`) to avoid some overhead in them.

* Decorate `get_nbytes` as `nogil`

Should avoid some extra overhead when calling this method. Also should
make it easier for the compiler to optimize and inline as needed.

* Add `get_nbytes_deps`

Provides a utility function for getting the total number of bytes of all
dependencies of this task. Should be useful in stealing where there is
a bottleneck computing this.

* Use `TaskState`'s `get_nbytes_deps` in stealing

Now that we have Cythonized this expression in this method, just use it
here to remove this bottleneck.