Provide `setup.py` option to profile Cython code
As it can be useful for development to profile the Cythonized code and
this requires an argument be provided to the extensions to enable, go
ahead and add some logic to `setup.py` to parse this option if provided.
If it is not provided, profiling is turned off, which is preferable as
this would impact the performance of the compiled code.