Create empty `SchedulerState` class
This class should ultimately handle everything related to task graph
state and performing transitions currently done in the `Scheduler`. This
is intended to be a base class of the `Scheduler`, which we will use to
manage related state. As this is a separate class that does not inherit
from other Python classes, it should be a good target for Cythonization.
This should allow us to more thoroughly optimize these components of the
`Scheduler`.  Afterwards the `Scheduler` should ideally be left with
communication part, `async` methods, user facing APIs, etc., used to
interact with other elements that are not really targets of the
Cythonization effort.