Use `dict` views onto `SortedDict` where possible
As Cython will leverage the Python C API on `dict` objects, it is faster
to use a `dict`. This works well with `SortedDict` as it is a subclass
of `dict`. However we have to keep in mind the `dict` will not have the
same order as the `SortedDict`. Also the `dict` won't handle
modifications correctly (unlike the `SortedDict`). So we need to be
mindful of how the object is being used before using the `dict` view.
That said, this works well in a lot of the code if order doesn't matter
as happens when things are placed in new `dict`s, `set`s, etc. or when
ordered objects like `list` end up being used in a way where order is
irrelevant like `sum`ming or where the iteration order is matched by
everything else. In cases where it is not obvious the `dict` can be
used, we simply skip it and stick with the `SortedDict`.