Fix some compile warnings
Use Cython's `compiled` flag to check whether the code is being
compiled. If `cython` can't even be `import`ed, then set the flag to
`False`. Then put all Cython bits in the `True` branch and all the
Python bits under the `False` branch. This should help Cython bracket
the Python fallbacks under the `False` case to optimize them out during
compilation (as opposed to leaving a bunch of unused fallback functions
around). As a result this should get rid of the unused symbol warnings
seen with the C compiler previously.