Drop custom `__eq__` from `Status` (#4270)
* Drop custom `__eq__` from `Status`

This appears to eat up a good chunk of time in places where the `Status`
needs to be checked. So try dropping this method in favor of the default
implementation.

* Drop tests that would be no longer needed

If we drop the special `__eq__` method, then these tests are no longer
relevant as they would error instead. So go ahead and drop them for now
should we decide to go ahead with this change.