Type annotation of `_reevaluate_occupancy_worker` (#4398)
* Return `total_duration` in `set_duration_estimate`

There is no need to retrieve this value when we already have it. So just
go ahead and return the variable directly.

* Group `old` and `new` variables

* Drop unused variable `nbytes`

* Use `.get(...)` to grab stealing extension

This way if it is `None`, we can check that and handle it much faster
(both in Python and Cython/C).

* Type iterated variable

* Assign `new - old` to `diff`

Hopefully should order computation to avoid a round off error.

* Type occupancy related variables as `double`

This should speed up mathematical operations with these variables.