Refactor `SchedulerState` from `Scheduler` (#4365)
* Move `worker_send` into transition functions

* Refactor `_task_to_msg` from `send_task_to_worker`

Provides a way for callers to simply construct the message if they are
not wanting to send it yet.

* Move `report` out of `_add_to_memory`

* Refactor out `_client_releases_keys`

This provides us a way to effectively call `client_releases_keys` from
other transitions without starting a new transition of its own.

* Collect client recs in `_add_to_memory`

* Use `_client_releases_keys` in transitions

* Refactor out `_task_to_report_msg`

Separates out the code needed to build a message for `report` based on
the `TaskState` in question from the actual call to `self.report`.

* Collect and send worker messages from transitions

* Handle `report` in `transition`

* Add method to send a message to a specific client

* Add `_task_to_client_msgs`

This converts a `TaskState` into a `dict` of messages with the keys
being the Clients to notify and the message being the report message.
Allows us to think of messages simply in terms of the message and where
it needs to be delivered without needing to know anything about the
`TaskState` it came from or the `ClientState`s involved.

* Replace `report_msg` with `client_msgs`

Instead of collecting a message to pass to `report` and letting the
relevant Clients be collected from the `TaskState` information later, go
ahead and collect that immediately while handling that `TaskState`.
These Clients then form the keys of `client_msgs` where the message
contains what was in `report_msg`. This allows us to keep all the
`TaskState` work contained to where it is relevant and can be handled
efficiently. Then the messaging out to Clients only needs be concerned
with the messages and where they go without needing to worry about what
they pertain to.

* Create empty `SchedulerState` class

This class should ultimately handle everything related to task graph
state and performing transitions currently done in the `Scheduler`. This
is intended to be a base class of the `Scheduler`, which we will use to
manage related state. As this is a separate class that does not inherit
from other Python classes, it should be a good target for Cythonization.
This should allow us to more thoroughly optimize these components of the
`Scheduler`.  Afterwards the `Scheduler` should ideally be left with
communication part, `async` methods, user facing APIs, etc., used to
interact with other elements that are not really targets of the
Cythonization effort.

* Move `transition*` methods into `SchedulerState`

Grabs all of the `transition` methods and methods they call and moves
them to `SchedulerState`.

* Add attributes for `SchedulerState`

Provides annotations for some attributes used by `SchedulerState`. Also
makes sure to provide an `_` prefix for these as they are accessible in
Cython only.

* Initialize attributes in `SchedulerState`

* Pass arguments to `super` class

* Use `SchedulerState` attributes

Now that we have typed attributes available through Cython. Use them
throughout the scheduler to improve performance. Also drop any duplicate
attribute assignments in `Scheduler`'s `__init__`.

* Use `cast` to access parent class attributes

Apparently Cython does not allow C-style access of attributes from
Python subclasses of extension types. However we can work around this by
`cast`ing `self` as the parent class type. To avoid muddying the code
too much, do this once per method and assign the result to `parent`.
Then use `parent` for all of the parent attribute access needed. Should
fix this issue while still allowing fast C-style attribute access in the
`Scheduler` subclass.

* Drop no longer needed `cast`s & local assignments

* Use `dict` views onto `SortedDict` where possible

As Cython will leverage the Python C API on `dict` objects, it is faster
to use a `dict`. This works well with `SortedDict` as it is a subclass
of `dict`. However we have to keep in mind the `dict` will not have the
same order as the `SortedDict`. Also the `dict` won't handle
modifications correctly (unlike the `SortedDict`). So we need to be
mindful of how the object is being used before using the `dict` view.
That said, this works well in a lot of the code if order doesn't matter
as happens when things are placed in new `dict`s, `set`s, etc. or when
ordered objects like `list` end up being used in a way where order is
irrelevant like `sum`ming or where the iteration order is matched by
everything else. In cases where it is not obvious the `dict` can be
used, we simply skip it and stick with the `SortedDict`.

* Add `@property`s for `SchedulerState` attributes

Ensures these Cython attributes can be retrieved by Python code.

* Take `worker_msgs` arg in `_propagate_forgotten`

Matches the behavior around `recommendations`. Also simplifies the
handling of `worker_msgs` (it's created only once).

* Use `parent` for private methods

As these will become `cfunc` decorated functions (only accessible from
the C API), we will need to `cast` to `SchedulerState` to get access to
them, which is what we do here in preparation for that.

* Annotate function arguments and return types

* Use `cfunc` on private `SchedulerState` methods

* Decorate `SchedulerState` methods with `@ccall`

Should allow faster C API calls to these methods in addition to their
existing Python APIs. Also disable exception checking when Python
objects are `return`ed since these already can and are checked for an
exception. Note that `@ccall` is not permitted on functions taking
`**kwargs` so those have been skipped.

* Add optional args to `transition_waiting_memory`

* Drop `**kwargs` from `_add_to_memory`

* Use `cfunc` to decorate `_add_to_memory`

* Make `SchedulerState` private methods functions

Gets them out of the `SchedulerState` virtual table so they can be used
like normal C functions. Should allow inlining and other optimizations
that apply.

* Collect `@property`s in `__pdict__`

As `@property`s are not placed in `__dict__`, but we need these in some
kind of `dict` for the dashboard. Collect them in `__pdict__`. This
should allow us to easily access them and place them in the dashboard.

https://stackoverflow.com/q/47432613/3877089

* Use `__pdict__` with `__dict__` in dashboard

Should give us the `@property`s as well.

* Refactor `consume_resources` & `release_resources`

* Refactor `aliases` into `SchedulerState`

* Refactor `coerce_hostname` into `SchedulerState`

* Refactor `task_metadata` into `SchedulerState`

* Refactor `remove_key` into `SchedulerState`