Just use `.get(...)` to retrieve transition func
Avoids checking for the presence of the key and then retrieving the
function corresponding the key by simply trying to get the function in
the first place or `None` if it is absent. As it is pretty quick to
check if something is `None` both in Python and Cython, this should
speed up the check and function retrieval time.