Better handling for dependencies that get re-added
A dependency might get created for the first time because it's required
by an added task.
Alternatively, a task may be added that requires a dependency that
already has a TaskState object, in which case we want to be more careful
about how we go about setting up dependency/dependent links and data
comms requests.