Transitions for scheduler reassignment following broken workers
Tasks can transition from `fetch` -> `waiting` or `flight` -> `waiting`
if the worker that has the data dies before the transfer is completed.

If that task is reassigned to a worker that was expecting just the data,
add the `runspec` and clear out any references to where the data _was_
coming from since it will now be executed locally.