Just check `_processing_on is not None`
We already know this is a `WorkerState` object or `None`. Also there is
not really a concept of a `False`y `WorkerState`. So the only way this
expression is `False` is if it is `None`. To speed up the check and
avoid unnecessary work. Just check that it `is not None`, which is a
fast pointer check.