Skip `stream.write` call for empty frames
Sometimes a frame winds up have `0` bytes. In these cases, make sure to
skip calling `stream.write` to avoid incurring the overhead. After all
it is trivial to create an empty frame on the receiving end as long as
we know it exists (the size of all frames are tracked) without involving
IO.