Transmit total frame size in initial message
This should allow us to allocate space for the entirety of the rest of
the message including the size of each frames and all following frames.
We can then unpack all of this information once received. By doing this
we are able to cutdown on addition send/recv calls that would otherwise
occur and spend less time in Tornado's IO handling.