Read & write all frames in one pass (#4506)
* Overwrite `frames` variable w/`join`ed result

Ensures that the separate `frames` are freed from memory before
proceeding to the next step of sending them out.

* Read all frames in one pass

Instead of doing multiple reads with `async`, just allocate one big
chunk of memory for all of the frames and read into it. Should cutdown
on the number of passes through Tornado needed to fill these frames.

* Transmit total frame size in initial message

This should allow us to allocate space for the entirety of the rest of
the message including the size of each frames and all following frames.
We can then unpack all of this information once received. By doing this
we are able to cutdown on addition send/recv calls that would otherwise
occur and spend less time in Tornado's IO handling.

* Handle `nframes` with rest of message

* Drop unneeded variable assignment

* Use packing and unpacking utility functions

Simplifies the code in the TCP path by leveraging existing utility
functions.

* Fix-up variable names

Make it a little easier to follow how the variables relate to each
other.

* Move packing/unpacking steps out of `try`

These are just binary serialization steps that are not really depended
on communication or issues that may come from sockets. So go ahead and
move them out of the `try` block.

* Group header building steps together

Also group `frames` and `frames_nbytes` steps together. Finally rewrites
the code to avoid use of constant for size of `"Q"`, which should make
it invariant to changes in that size.

* Use `struct.calcsize` to get size

Should avoid issues on platforms where this may not be the exact size.

* Concatenate small frames beforehand

To simplify the logic, just concatenate small frames before doing any
sends. This way we can use the same code path for all sends.

* Use trick to enqueue frames to write beforehand

Tornado has an internal queue that uses to hold frames before writing
them. To avoid needing to track and wait on various `Future`s and the
amount of data sent, we can just enqueue all of the frames we want to
send before a send even happens and then start the write. This way
Tornado already has all of the data we plan to send once it starts
working. In the meantime, we are able to carry on with other tasks while
this gets handled in the background.

https://github.com/tornadoweb/tornado/blob/6cdf82e927d962290165ba7c4cccb3e974b541c3/tornado/iostream.py#L537-L538